fixed a bug

+++ Lab Assignment #4Linear Equations/1.c
@@ -7,19 +7,19 @@ int main()
   int n;
   printf("Enter the dimension of system of equation: ");
   scanf("%d", &n);
-  float matrixA[n][n], matrixB[n];
+  double matrixA[n][n], matrixB[n];
   printf("Enter the coefficients row-wise:\n");
   for (int i = 0; i < n; i++)
   {
     for (int j = 0; j < n; j++)
     {
-      scanf("%f", &matrixA[i][j]);
+      scanf("%lf", &matrixA[i][j]);
     }
   }
   printf("Enter the RHS vector: ");
   for (int i = 0; i < n; i++) 
   {
-    scanf("%f", &matrixB[i]);
+    scanf("%lf", &matrixB[i]);
   }
 
 printf("\nMatrix A: \n");
@@ -27,21 +27,21 @@ for (int i = 0; i < n; i++)
 {
   for (int j = 0; j < n; j++) 
   {
-    printf("%f\t", matrixA[i][j]);
+    printf("%lf\t", matrixA[i][j]);
   }
   printf("\n");
 }
 printf("\nMatrix B: \n");
 for (int i = 0; i < n; i++)
 {
-  printf("%f\n", matrixB[i]);
+  printf("%lf\n", matrixB[i]);
 }
   for (int i = 0; i < n; i++)
   {
-    float pivot = matrixA[i][i];
+    double pivot = matrixA[i][i];
     for (int j = i + 1; j < n; j++)
     {
-      float constant = matrixA[j][i] / pivot;
+      double constant = matrixA[j][i] / pivot;
       for (int k = i; k < n; k++)
       {
         matrixA[j][k] = matrixA[j][k] - constant * matrixA[i][k];
@@ -59,19 +59,19 @@ for (int i = 0; i < n; i++)
   {
     for (int j = 0; j < n; j++)
     {
-      printf("%f\t", matrixA[i][j]);
+      printf("%lf\t", matrixA[i][j]);
     }
     printf("\n");
   }
   printf("\nMatrixB:\n");
   for (int i = 0; i < n; i++) 
   {
-    printf("%f\n", matrixB[i]);
+    printf("%lf\n", matrixB[i]);
   }
-  float matrixX[n];
+  double matrixX[n];
   for (int i = n - 1; i >= 0; i--) 
   {
-    float sum = 0;
+    double sum = 0;
     for (int j = i + 1; j <= n; j++) 
     {
       sum = sum + matrixA[i][j] * matrixX[j];
@@ -80,7 +80,7 @@ for (int i = 0; i < n; i++)
   }
   printf("\nSolution X: \n");
   for (int i = 0; i < n; i++) {
-    printf("%f\n", matrixX[i]);
+    printf("%lf\n", matrixX[i]);
   }
   return 0;
 }
